moved set_expired_to_archived from scheduled_tasks to commands
