+++ dds_web/__init__.py
@@ -266,12 +266,16 @@ def create_app(testing=False, database_uri=None):
             create_new_unit,
             update_uploaded_file_with_log,
             lost_files_s3_db,
+            set_available_to_expired,
         )
 
         app.cli.add_command(fill_db_wrapper)
         app.cli.add_command(create_new_unit)
         app.cli.add_command(update_uploaded_file_with_log)
         app.cli.add_command(lost_files_s3_db)
+
+        # Add flask commands - cronjobs
+        app.cli.add_command(set_available_to_expired)
 
         # Make version available inside jinja templates:
         @app.template_filter("dds_version")
